requirements.txt
absl-py==0.13.0
chex==0.0.8
contextlib2==21.6.0
dill==0.3.4
dm-env==1.5
dm-haiku==0.0.4
dm-tree==0.1.6
flatbuffers==2.0
jax==0.2.19
jaxlib==0.1.70
ml-collections==0.1.0
numpy==1.21.2
opt-einsum==3.3.0
optax==0.0.9
PyYAML==5.4.1
scipy==1.7.1
six==1.16.0
tabulate==0.8.9
toolz==0.11.1
